requirements.txt
Pillow
ImageHash
Werkzeug==1.0.1
Flask